+++ core/src/feature/moment.c
@@ -58,7 +58,7 @@ int compute_2nd_moment(const float *pic, int w, int h, int stride, double *score
     for (int i = 0; i < h; ++i) {
         for (int j = 0; j < w; ++j) {
             pic_ = pic[i * stride_ + j];
-            cum += pic_ * pic_;
+            cum += (double)pic_ * pic_;
         }
     }
     cum /= ((double)w * h);
